+++ arrow-udf-wasm/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update `wasmtime` to v19.
+- Update `arrow` version to >=50.
+
 ## [0.2.0] - 2024-02-29
 
 ### Added