+++ arrow-udf-wasm/CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.2] - 2024-04-25
+
+### Added
+
+- Support arrow-udf v0.3.
+- Add `Runtime::abi_version`.
+
+### Changed
+
+- Update `wasmtime` to v20.
+
 ## [0.2.1] - 2024-04-07
 
 ### Changed